+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Refactor certificate verification and implement UDID-cert authentication (#358, #429)
 * Cleanup DEP library and integrate into main project (#504, #505)
 * Add API endpoint to retrieve APNS certificate (#503)
+* Remove deprecated `-apns` flags from server startup (#528)
+* Move API calls to list endpoints from HTTP GET to HTTP POST (#522, #523, #524, #525, #526)
 
 ## [v1.4.0](https://github.com/micromdm/micromdm/compare/v1.3.1...v1.4.0) September 6 2018